+++ src/renderer/state.ts
@@ -673,6 +673,7 @@ export interface SessionState {
     workspacePanelMode: 'resources' | 'steps';
 
     showHumanReadableDataSourceTitles: boolean;
+    showLayerTextOverlay: boolean;
 
     /**
      * backendConfig settings are applied on restart (of the WebAPI) only.